fix some build errors & add some additional checks to compiler

+++ src/lib/sorting.h
@@ -12,11 +12,11 @@
 // Using radix sort can become too verbose in some string-heavy places, so another sorting algorithm may be introduced
 
 typedef struct SortEntry {
-    u64 key;   // value that array needs to be sorted around. Floating-point needs special handling, see functions below
-    u64 value; // any value that user can use to use sort data
+    u32 key;   // value that array needs to be sorted around. Floating-point needs special handling, see functions below
+    u32 value; // any value that user can use to use sort data
 } SortEntry;
 // Construct soring key from floating-pointvalue
-u64 f64_to_sort_key(f64 value);
+u32 f64_to_sort_key(f64 value);
 
 // O(n)
 // Sorted result is stored in entries
